Refactoring to introduce custom processing
